+++ doc-examples/openhiven_test.py
@@ -22,5 +22,6 @@ async def on_ready():
 async def on_message_create():
     print(f"Message was created")
 
+
 if __name__ == '__main__':
-    client.run("Insert token", )
+    client.run("y9JvliBe8fbOmmSAmS09sHvWk6FVV2ZdOtVenZAsoN0xH3OwXmJZIooTh3PuKyECp0UUDDfuS6gT6MbaBDYecryEky4HS2LMq1Sp3WM5u7jakITMON5QXk9NseiL3Syr", )